app/requirements.txt
Django>=3.2,<4.0
django-cors-headers>=3.7.0
django-rest-knox>=4.1.0
djangorestframework>=3.12.4,<4.0
gunicorn>=20.1.0
Pillow>=8.2.0,<9.0
psycopg2-binary>=2.8.6
celery>=5.1.0
redis>=3.5.3
requests==2.25.1
